% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cycle_net.R
\name{get_cycle_net}
\alias{get_cycle_net}
\title{Get Fortaleza cycle network from MyMaps}
\usage{
get_cycle_net(forceDownload = FALSE)
}
\arguments{
\item{forceDownload}{A logical. When TRUE force download of the kml from google. Default is FALSE, in this case try use cache kml downloaded if available.}
}
\value{
An object of class `sf`, which includes the mymaps columns and a additional column 'layer'.
}
\description{
`get_cycle_net` download kml objects from mymaps and read into R.
}
\details{
- geometries downloaded from 'https://bit.ly/ciclofortaleza'
}
\examples{

net <- get_cycle_net()


}
