% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_future_pkgs.R
\name{ld_future_pkgs}
\alias{ld_future_pkgs}
\title{Load Futureverse Packages}
\usage{
ld_future_pkgs()
}
\value{
`NULL` (the function is used for its side effect of loading packages)
}
\description{
`ld_future_pkgs` loads all packages from the `futureverse` ecosystem in the current R session.
}
\details{
This function retrieves the list of packages from `futureverse::futureverse_packages()` and loads them using `library()` with the `character.only = TRUE` option.
}
\examples{
# Load all futureverse packages
ld_future_pkgs()
}
